# sdd/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 任务管理系统 API
  description: |
    任务管理系统后端 API 文档
    
    核心功能：
    - Excel 导入任务数据
    - 模板管理
    - Rundown 发布清单管理
    - 任务执行控制
    - Jenkins/Ansible 自动化集成
    - 定时任务调度
    - 批量执行
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发服务器
  - url: http://{host}:{port}/api/v1
    description: 生产服务器
    variables:
      host:
        default: localhost
      port:
        default: "8000"
...
paths:
...
  /rundowns:
    get:
      tags:
        - Rundowns
      summary: 获取 Rundown 列表
      operationId: listRundowns
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, partial_success, success, aborted, failed]
        - name: template_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rundown 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRundowns'

    post:
      tags:
        - Rundowns
      summary: 创建空白 Rundown
      operationId: createRundown
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRundownRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rundown'
  /rundowns/{rundown_id}:
    get:
      tags:
        - Rundowns
      summary: 获取 Rundown 详情
      operationId: getRundown
      parameters:
        - name: rundown_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rundown 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RundownDetail'

    put:
      tags:
        - Rundowns
      summary: 更新 Rundown
      operationId: updateRundown
      parameters:
        - name: rundown_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRundownRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rundown'

    delete:
      tags:
        - Rundowns
      summary: 删除 Rundown
      operationId: deleteRundown
      parameters:
        - name: rundown_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功
...
  /rundowns/{rundown_id}/tasks:
    get:
      tags:
        - Tasks
      summary: 获取 Rundown 下的任务列表
      operationId: listTasks
      parameters:
        - name: rundown_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 任务列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      tags:
        - Tasks
      summary: 添加任务到 Rundown
      operationId: addTask
      parameters:
        - name: rundown_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{task_id}:
    get:
      tags:
        - Tasks
      summary: 获取任务详情
      operationId: getTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'

    put:
      tags:
        - Tasks
      summary: 更新任务
      operationId: updateTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

    delete:
      tags:
        - Tasks
      summary: 删除任务
      operationId: deleteTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功
...
  /tasks/{task_id}/reorder:
    put:
      tags:
        - Tasks
      summary: 调整任务顺序
      operationId: reorderTasks
      parameters:
        - name: rundown_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - task_ids
              properties:
                task_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: 按顺序排列的任务ID列表
      responses:
        '200':
          description: 排序更新成功

  # ==================== Scheduling ====================
...
  /config:
    get:
      tags:
        - Config
      summary: 获取系统配置列表
      operationId: listConfig
      responses:
        '200':
          description: 配置列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SystemConfig'

    post:
      tags:
        - Config
      summary: 创建/更新系统配置
      operationId: setConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetConfigRequest'
      responses:
        '200':
          description: 设置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfig'

# ==================== Components ====================
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
...
    PaginatedRundowns:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Rundown'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    Rundown:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rundown_code:
          type: string
        name:
          type: string
        description:
          type: string
        template_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, partial_success, success, aborted, failed]
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
    RundownDetail:
      allOf:
        - $ref: '#/components/schemas/Rundown'
        - type: object
          properties:
            tasks:
              type: array
              items:
                $ref: '#/components/schemas/Task'
    CreateRundownRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    UpdateRundownRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    # Tasks
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rundown_id:
          type: string
          format: uuid
        name:
          type: string
        task_type:
          type: string
          enum: [jenkins, ansible]
        config:
          type: object
        order_index:
          type: integer
        status:
          type: string
          enum: [pending, running, success, failed, aborted, timeout]
        external_id:
          type: string
        external_url:
          type: string
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
    TaskDetail:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            execution_logs:
              type: array
              items:
                $ref: '#/components/schemas/ExecutionLog'
    CreateTaskRequest:
      type: object
      required:
        - name
        - task_type
        - config
      properties:
        name:
          type: string
        task_type:
          type: string
          enum: [jenkins, ansible]
        config:
          type: object
        order_index:
          type: integer
    UpdateTaskRequest:
      type: object
      properties:
        name:
          type: string
        task_type:
          type: string
          enum: [jenkins, ansible]
        config:
          type: object
    ExecutionLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        execution_id:
          type: string
        status:
          type: string
        output:
          type: string
        error_output:
          type: string
        duration_ms:
          type: integer
        created_at:
          type: string
          format: date-time

    # Execution
...
    SystemConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        config_key:
          type: string
        value_type:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SetConfigRequest:
      type: object
      required:
        - config_key
        - config_value
      properties:
        config_key:
          type: string
        config_value:
          type: string
        value_type:
          type: string
          default: string
        description:
          type: string

    # Error
...